refactor : 폴더 명 통일 및 분산되어 있던 파일들 통합 [#66]
