+++ frontend/src/routes/chat.tsx
@@ -1,4 +1,4 @@
-import { Chat } from "@/pages/chat";
+import { Chat } from "@/features/chat";
 import { createFileRoute } from "@tanstack/react-router";
 
 export const Route = createFileRoute("/chat")({
