+++ frontend/src/routes/login.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from "@tanstack/react-router";
-import { PredictDetail } from "@/pages/predict-detail";
+import { PredictDetail } from "@/features/predict-detail";
 
 export const Route = createFileRoute("/login")({
   component: PredictDetail,
