+++ frontend/src/routes/my-page.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from "@tanstack/react-router";
-import { BettingPredictResult } from "@/pages/betting-predict-result";
+import { BettingPredictResult } from "@/features/betting-predict-result";
 
 export const Route = createFileRoute("/my-page")({
   component: RouteComponent,
