+++ frontend/src/routes/waiting-room.tsx
@@ -1,4 +1,4 @@
-import { WaitingRoom } from "@/pages/WaitingRoom";
+import { WaitingRoom } from "@/features/waiting-room";
 import { createFileRoute } from "@tanstack/react-router";
 
 export const Route = createFileRoute("/waiting-room")({
